+++ HabboHotel/Moderation/IModerationTicketService.cs
@@ -5,7 +5,7 @@ namespace Plus.HabboHotel.Moderation;
 public interface IModerationTicketService
 {
     Task SendOpenState(GameClient session);
-    Task Submit(GameClient session, string message, int category, int reportedUserId, int type, IReadOnlyCollection<ModerationTicketChatEntry> reportedChats);
+    Task Submit(GameClient session, string message, int category, int reportedUserId, int type, IReadOnlyCollection<ModerationTicketChatEntry> reportedChats, ModerationTicketContext? context = null);
     Task Close(GameClient session, int result, int ticketId);
     Task Pick(GameClient session, int ticketId);
     Task Release(GameClient session, IReadOnlyCollection<int> ticketIds);